fix: refactoring class names
